fix: Incorrect package name for ConceptDescriptionRepository-Backend (#611)
