+++ basyx.conceptdescriptionrepository/basyx.conceptdescriptionrepository-backend-inmemory/src/test/java/org/eclipse/digitaltwin/basyx/conceptdescriptionrepository/backend/TestInMemoryConceptDescriptionRepository.java
@@ -23,18 +23,15 @@
  * SPDX-License-Identifier: MIT
  ******************************************************************************/
 
+package org.eclipse.digitaltwin.basyx.conceptdescriptionrepository.backend;
 
-package org.eclipse.digitaltwin.basyx.conceptdescriptionrepository;
-
-import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.*;
 
 import java.util.Arrays;
 import java.util.Collection;
 
 import org.eclipse.digitaltwin.aas4j.v3.model.ConceptDescription;
-import org.eclipse.digitaltwin.basyx.aasrepository.backend.ConceptDescriptionBackendProvider;
-import org.eclipse.digitaltwin.basyx.aasrepository.backend.CrudConceptDescriptionRepository;
-import org.eclipse.digitaltwin.basyx.aasrepository.backend.SimpleConceptDescriptionRepositoryFactory;
+import org.eclipse.digitaltwin.basyx.conceptdescriptionrepository.ConceptDescriptionRepository;
 import org.eclipse.digitaltwin.basyx.conceptdescriptionrepository.core.ConceptDescriptionRepositorySuite;
 import org.eclipse.digitaltwin.basyx.conceptdescriptionrepository.core.DummyConceptDescriptionFactory;
 import org.eclipse.digitaltwin.basyx.core.exceptions.CollidingIdentifierException;
@@ -47,9 +44,9 @@ import org.junit.Test;
  *
  */
 public class TestInMemoryConceptDescriptionRepository extends ConceptDescriptionRepositorySuite {
-	
+
 	private static final String CONFIGURED_CD_REPO_NAME = "configured-cd-repo-name";
-	
+
 	private ConceptDescriptionBackendProvider backendProvider = new ConceptDescriptionInMemoryBackendProvider();
 
 	@Override
@@ -61,18 +58,18 @@ public class TestInMemoryConceptDescriptionRepository extends ConceptDescription
 	protected ConceptDescriptionRepository getConceptDescriptionRepository(Collection<ConceptDescription> conceptDescriptions) {
 		return new SimpleConceptDescriptionRepositoryFactory(backendProvider, conceptDescriptions).create();
 	}
-	
+
 	@Test
-    public void getConfiguredInMemoryConceptDescriptionRepositoryName() {
+	public void getConfiguredInMemoryConceptDescriptionRepositoryName() {
 		ConceptDescriptionRepository repo = new CrudConceptDescriptionRepository(backendProvider, CONFIGURED_CD_REPO_NAME);
-		
+
 		assertEquals(CONFIGURED_CD_REPO_NAME, repo.getName());
 	}
 
 	@Test(expected = CollidingIdentifierException.class)
 	public void idCollisionDuringConstruction() {
 		Collection<ConceptDescription> conceptDescriptionsWithCollidingIds = createConceptDescriptionCollectionWithCollidingIds();
-		
+
 		new CrudConceptDescriptionRepository(backendProvider, conceptDescriptionsWithCollidingIds);
 	}
